[feat] FAQ 관리 기능 > 토큰검사 로직 추가
- 추가/수정/삭제 토큰 검사 로직 추가
- 기존 session 처리 로직 삭제
- 삭제 > DELETE Http method 사용

+++ _admin/admin_faq.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, Request, Form
+from fastapi import APIRouter, Depends, Form, HTTPException, Request
 from fastapi.responses import RedirectResponse
 from fastapi.templating import Jinja2Templates
 from sqlalchemy.orm import Session
@@ -6,11 +6,14 @@ from database import get_db
 import models
 from common import *
 
+
 router = APIRouter()
 templates = Jinja2Templates(directory=TEMPLATES_DIR)
 # 파이썬 함수 및 변수를 jinja2 에서 사용할 수 있도록 등록
 # TODO: admin/base.html에서만 사용하는 함수이므로 이동이 필요함
 templates.env.globals["get_admin_menus"] = get_admin_menus
+# TODO: form에서 공용으로 사용하는 함수이므로 이동이 필요함
+templates.env.globals["generate_one_time_token"] = generate_one_time_token
 
 
 @router.get("/faq_master_list")
@@ -20,6 +23,7 @@ def faq_master_list(request: Request, db: Session = Depends(get_db)):
     request.session["menu_key"] = "300700"
 
     faq_masters = db.query(model).order_by(model.fm_order).all()
+
     return templates.TemplateResponse(
         "admin/faq_master_list.html", {"request": request, "faq_masters": faq_masters}
     )
@@ -28,11 +32,9 @@ def faq_master_list(request: Request, db: Session = Depends(get_db)):
 @router.get("/faq_master_form")
 def faq_master_add_form(request: Request):
     """FAQ관리 등록 폼"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
 
     return templates.TemplateResponse(
-        "admin/faq_master_form.html", {"request": request, "faq_master": None, "token": token}
+        "admin/faq_master_form.html", {"request": request, "faq_master": None}
     )
 
 
@@ -49,8 +51,9 @@ def faq_master_add(
         fm_order: int = Form(0)
     ):
     """FAQ관리 등록 처리"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
+    # 토큰 검사
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
 
     faq_master = models.FaqMaster(
         fm_subject=fm_subject
@@ -69,13 +72,10 @@ def faq_master_add(
 @router.get("/faq_master_form/{fm_id}")
 def faq_master_update_form(fm_id: int, request: Request, db: Session = Depends(get_db)):
     """FAQ관리 수정 폼"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
-
     faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
 
     return templates.TemplateResponse(
-        "admin/faq_master_form.html", {"request": request, "faq_master": faq_master, "token": token}
+        "admin/faq_master_form.html", {"request": request, "faq_master": faq_master}
     )
 
 
@@ -93,8 +93,9 @@ def faq_master_update(
         fm_order: int = Form(0)
     ):
     """FAQ관리 수정 처리"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
+    # 토큰 검사
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
 
     faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
 
@@ -108,19 +109,25 @@ def faq_master_update(
 
     return RedirectResponse(f"/admin/faq_master_form/{faq_master.fm_id}", status_code=303)
 
-# faq_master 삭제처리 함수 구현해줘
-@router.get("/faq_master_form/{fm_id}/delete")
+
+@router.delete("/faq_master_form/{fm_id}/delete")
 def faq_master_delete(
         fm_id: int,
         request: Request,
         db: Session = Depends(get_db),
     ):
     """FAQ관리 삭제 처리"""
+    # 토큰 검사
+    # DELETE 요청일 경우, Body에 토큰이 없으므로 쿼리스트링에서 토큰을 얻는다.
+    token = request.query_params.get("token")
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
+    
     faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
     db.delete(faq_master)
     db.commit()
 
-    return RedirectResponse(f"/admin/faq_master_list", status_code=303)
+    return {"message": "FAQ가 성공적으로 삭제되었습니다."}
 
 
 @router.get("/faq_master_form/{fm_id}/list")
@@ -141,13 +148,10 @@ def faq_list(fm_id: int, request: Request, db: Session = Depends(get_db)):
 @router.get("/faq_master_form/{fm_id}/add")
 def faq_add_form(fm_id: int, request: Request, db: Session = Depends(get_db)):
     """FAQ관리 등록 폼"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
-
     faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
 
     return templates.TemplateResponse(
-        "admin/faq_form.html", {"request": request, "faq_master": faq_master, "faq": None, "token": token}
+        "admin/faq_form.html", {"request": request, "faq_master": faq_master, "faq": None}
     )
 
 
@@ -162,8 +166,9 @@ def faq_add(
         fa_content: str = Form(...),
     ):
     """FAQ관리 등록 처리"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
+    # 토큰 검사
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
 
     faq = models.Faq(
         fm_id=fm_id,
@@ -180,14 +185,11 @@ def faq_add(
 @router.get("/faq_master_form/{fm_id}/faq/{fa_id}")
 def faq_update_form(fa_id: int, request: Request, db: Session = Depends(get_db)):
     """FAQ관리 등록 폼"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
-
     faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
     faq_master = faq.faq_master
 
     return templates.TemplateResponse(
-        "admin/faq_form.html", {"request": request, "faq_master": faq_master, "faq": faq, "token": token}
+        "admin/faq_form.html", {"request": request, "faq_master": faq_master, "faq": faq}
     )
 
 
@@ -203,8 +205,9 @@ def faq_update(
         fa_content: str = Form(...),
     ):
     """FAQ관리 등록 처리"""
-    token = hash_password(hash_password(""))  # 토큰값을 아무도 알수 없게 만듬
-    request.session["token"] = token
+    # 토큰 검사
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
 
     faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
 
@@ -216,17 +219,21 @@ def faq_update(
     return RedirectResponse(f"/admin/faq_master_form/{fm_id}/faq/{faq.fa_id}", status_code=303)
 
 
-@router.get("/faq_master_form/{fm_id}/faq/{fa_id}/delete")
-def faq_delete(
-        fm_id: int,
+@router.delete("/faq_master_form/{fm_id}/faq/{fa_id}")
+async def faq_delete(
         fa_id: int,
         request: Request,
         db: Session = Depends(get_db),
     ):
     """FAQ 항목 삭제 처리"""
+    # 토큰 검사
+    # DELETE 요청일 경우, Body에 토큰이 없으므로 쿼리스트링에서 토큰을 얻는다.
+    token = request.query_params.get("token")
+    if not validate_one_time_token(token):
+        raise HTTPException(status_code=403, detail="Invalid token.")
+
     faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
     db.delete(faq)
     db.commit()
 
-    return RedirectResponse(f"/admin/faq_master_form/{fm_id}/list", status_code=303)
-    
+    return {"message": "FAQ 항목이 성공적으로 삭제되었습니다."}
